Make form model fragment static
After upgrading support library, app crashes, complaining form fragment is not static.
Make form fragment static (still retain instance) and pass form object via `newInstance`.
