+++ androidvalidatedforms/src/main/java/com/azavea/androidvalidatedforms/FormWithAppCompatActivity.java
@@ -15,9 +15,9 @@ import android.view.ViewGroup;
 import android.view.WindowManager.LayoutParams;
 
 import com.azavea.androidvalidatedforms.controllers.ImageController;
+import com.azavea.androidvalidatedforms.FormModelEnclosure.FormModel;
 import com.azavea.androidvalidatedforms.tasks.DisplayFormTask;
 
-import java.io.File;
 import java.lang.ref.WeakReference;
 import java.util.HashMap;
